+++ feature/home/src/main/res/values/strings.xml
@@ -28,4 +28,8 @@
     <string name="feature_home_innerview_question_list_title">질문 목록</string>
     <string name="feature_home_innerview_question_description">인터뷰 전체 재생 시, 질문 목록의 순서에 따라 재생됩니다.</string>
     <string name="feature_home_innerview_detail_description_recording">촬영중인 인터뷰가 있습니다</string>
+
+    <string name="feature_home_permission_dialog_title">권한 설정</string>
+    <string name="feature_home_permission_dialog_content">알림을 허용하면 새로운 인터뷰 촬영이 가능할 때 알림을 받을 수 있습니다. \n허용하시겠습니까?</string>
+    <string name="feature_home_permission_dialog_approve">허용</string>
 </resources>
