+++ feature/main/src/main/java/com/dev/innerview/feature/main/MainScreen.kt
@@ -10,7 +10,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import com.dev.innerview.feature.main.component.MainBottomBar
 import com.dev.innerview.feature.main.component.MainNavHost
-import com.dev.innerview.feature.home.NotificationPermissionLauncher
 import kotlinx.collections.immutable.toPersistentList
 import kotlinx.coroutines.launch
 
@@ -41,8 +40,6 @@ internal fun MainScreen(
         onShowErrorSnackBar = onShowErrorSnackBar,
         snackBarHostState = snackBarHostState
     )
-
-    NotificationPermissionLauncher()
 }
 
 @Composable
